requirements.txt
Flask==2.0.1
pymongo==4.0.1
python-dotenv==0.19.1
gunicorn==20.1.0
werkzeug==2.3.7
pymongo[srv]
jinja2==3.0.3
MarkupSafe==2.0.1